修改了编辑server权限bug，及删除和新增加入了权限控制

+++ api/views/server.py
@@ -2,17 +2,18 @@ from django.shortcuts import render
 
 # Create your views here.
 from api.libs.base import CoreView
-from cmdb.models import Asset, BusinessUnit, Cabinet, Tags
+from cmdb.models import Asset, BusinessUnit, Cabinet, Tags, EventLog
 from django.db.models import Q
 from api.libs.cmdb_agent import CmdbCollector
 from api.libs.asset_handler import AssetHandler
 from guardian.shortcuts import get_objects_for_user
 from django.contrib.auth.models import User
 from account.models import UserProfile
+from django.conf import settings
 
 
 class Server(CoreView):
-    login_required_action = ["get_list"]
+    login_required_action = ["get_list", "post_create", "get_detail", "post_delete", "post_change"]
 
     # permission_view_map = {
     #     "get_list": "can_view_asset"
@@ -41,6 +42,11 @@ class Server(CoreView):
         project_id = self.parameters('project')
         ipaddresses = self.parameters('ipaddresses').split(';')
         port = self.parameters('port')
+        project_obj = BusinessUnit.objects.filter(id=project_id).first()
+        if not project_obj or not self.request.user.has_perm('add_project_asset', project_obj):
+            self.response_data["status"] = False
+            self.status_code = 403
+            return
         for ipaddress in ipaddresses:
             cmdb_collector = CmdbCollector(host=ipaddress, username=username, password=password, ssh_port=port)
             cmdb_collector.collector_all()
@@ -78,12 +84,16 @@ class Server(CoreView):
 
     def post_delete(self):
         asset_id = self.parameters("server_id")
-        try:
-            asset_obj = Asset.objects.filter(id=asset_id).first()
-            asset_obj.delete()
-        except Exception as e:
+        asset_obj = Asset.objects.filter(id=asset_id).first()
+        if asset_obj:
+            if self.request.user.has_perm("del_project_asset", asset_obj.business_unit):
+                asset_obj.delete()
+            else:
+                self.response_data['status'] = False
+                self.status_code = 403
+        else:
             self.response_data['status'] = False
-            self.status_code = 500
+            self.status_code = 404
 
     def post_change(self):
         asset_id = self.parameters("id")
@@ -92,38 +102,70 @@ class Server(CoreView):
         if value == "":
             value = None
         asset_obj = Asset.objects.filter(id=asset_id).first()
+        filed_name_map = {
+            "asset_num": "资产编号",
+            "status": "运营状态",
+            "management_ip": "管理IP",
+            "raid_type": "Raid类型",
+            "price": "价格",
+            "trade_date": "购买时间",
+            "expire_date": "过保时间",
+            "cabinet": "机柜",
+            "admin": "负责人",
+            "operation": "运维人员",
+            "business_unit": "所属项目",
+        }
+        status_map = dict()
+        for status_item in settings.ASSET_STATUS_CHOICES:
+            status_map[status_item[0]] = status_item[1]
         if asset_obj:
             if not self.request.user.has_perm("change_project_asset", asset_obj.business_unit):
                 self.response_data["status"] = False
                 self.status_code = 403
+                return
             filed_name = filed_name.split(".")
+            old_value = None
+            new_value = None
+
             if filed_name[0] == 'base':
                 if filed_name[1] not in ["admin", "raid_type", "cabinet", "business_unit", "operation", "tags"]:
-                    setattr(asset_obj, filed_name[1], value)
-                    asset_obj.save()
+                    old_value = getattr(asset_obj, filed_name[1])
+                    if not value == old_value:
+                        setattr(asset_obj, filed_name[1], value)
+                        asset_obj.save()
+                        if filed_name[1] == "status":
+                            old_value = status_map[old_value]
+                            new_value = status_map[int(value)]
+                        else:
+                            new_value = value
                 if filed_name[1] == "cabinet":
+                    old_value = asset_obj.cabinet.number if asset_obj.cabinet else "空"
                     cabinet_obj = Cabinet.objects.filter(id=value).first()
-                    asset_obj.cabinet = cabinet_obj
-                    asset_obj.save()
+                    new_value = cabinet_obj.number if cabinet_obj else "空"
+                    if not old_value == new_value:
+                        asset_obj.cabinet = cabinet_obj
+                        asset_obj.save()
                 if filed_name[1] == "business_unit":
+                    old_value = asset_obj.business_unit.name if asset_obj.business_unit else "空"
                     business_unit_obj = BusinessUnit.objects.filter(id=value).first()
-                    asset_obj.business_unit = business_unit_obj
-                    asset_obj.save()
+                    new_value = business_unit_obj.name if business_unit_obj else "空"
+                    if not old_value == new_value:
+                        asset_obj.business_unit = business_unit_obj
+                        asset_obj.save()
                 if filed_name[1] == "admin":
+                    old_value = asset_obj.admin.username if asset_obj.admin else "空"
                     admin_obj = UserProfile.objects.filter(id=value).first()
-                    if admin_obj:
-                        asset_obj.admin = admin_obj.user
+                    new_value = admin_obj.user.username if admin_obj and admin_obj.user else "空"
+                    if not old_value == new_value:
+                        asset_obj.admin = admin_obj.user if admin_obj else None
                         asset_obj.save()
-                    else:
-                        asset_obj.admin = None
-                        asset_obj.save()
+
                 if filed_name[1] == "operation":
+                    old_value = asset_obj.operation.username if asset_obj.operation else "空"
                     operation_obj = UserProfile.objects.filter(id=value).first()
-                    if operation_obj:
-                        asset_obj.operation = operation_obj.user
-                        asset_obj.save()
-                    else:
-                        asset_obj.operation = None
+                    new_value = operation_obj.user.username if operation_obj and operation_obj.user else "空"
+                    if not old_value == new_value:
+                        asset_obj.operation = operation_obj.user if operation_obj else None
                         asset_obj.save()
                 if filed_name[1] == "tags":
                     asset_obj.tags = []
@@ -143,7 +185,15 @@ class Server(CoreView):
                 setattr(asset_obj.server, filed_name[1], value)
                 asset_obj.server.save()
 
-
+            if new_value or not old_value:
+                msg = "变更%s：旧配置：%s 新配置：%s" % (filed_name_map[filed_name[1]], old_value, new_value)
+                log_obj = EventLog(asset=asset_obj,
+                                   event_type=1,
+                                   operater=self.request.user,
+                                   component=None,
+                                   detail=msg
+                                   )
+                log_obj.save()
         else:
             self.response_data['status'] = False
             self.status_code = 404
